// types.ts
export interface SelectedNetworkOption {
  label: string;
  value: number;
}

export interface SafeDappInfo {
  id: number;
  url: string;
  name: string;
  iconUrl: string;
}

export interface TxnDataType {
  id: number;
  from: string;
  to: string;
  data: string;
  value: string;
}